Updated loginitem to work with current cleanup standards
