+++ ttps/persistence/macos/loginitem/README.md
@@ -20,7 +20,7 @@ the specified command or script runs upon user login.
 Add a specific script as a Login Item using Swift:
 
 ```bash
-ttpforge run ttps/persistence/macos/login-item/loginitem.yaml \
+ttpforge run ttps/persistence/macos/loginitem/loginitem.yaml \
   --arg command_or_path="/Users/Shared/myScript.sh" \
   --arg cleanup=true
 ```
@@ -28,7 +28,7 @@ ttpforge run ttps/persistence/macos/login-item/loginitem.yaml \
 Alternatively, run a direct command:
 
 ```bash
-ttpforge run ttps/persistence/macos/login-item/loginitem.yaml \
+ttpforge run ttps/persistence/macos/loginitem/loginitem.yaml \
   --arg command_or_path="osascript -e 'display dialog \"Hello World\"'" \
   --arg cleanup=true
 ```
@@ -40,7 +40,7 @@ ttpforge run ttps/persistence/macos/login-item/loginitem.yaml \
   - Execute the compiled binary. It adds the given command or script to the
     macOS Login Items list, ensuring its execution every time the user logs
     in.
-  - If `cleanup` is set to true, it will:
+  - It will then run the following cleanup steps:
     1. Delay for 15 seconds before initiating the cleanup process.
     1. Remove the login item and associated files.
     1. Revert the persistence.
@@ -55,9 +55,6 @@ TTP. Upon execution:
   shell script that opens the Calculator app and sets the appropriate permissions.
 - It then attempts to add the provided command or path to the macOS Login
   Items list using JavaScript for Automation (JXA).
-- If the cleanup flag is true, it will remove the Login Item using a
-  combination of Apple's CoreServices and Security framework functions, as
-  well as the SystemConfiguration framework to manage the list of Login Items.
 - For the default behavior where the path is `/Users/Shared/calc.sh`, it
   will also delete the created shell script from the filesystem.
