+++ ttps/persistence/macos/loginitem/loginitem.swift
@@ -2,59 +2,18 @@ import Foundation
 import Cocoa
 import OSAKit
 
-// Function for Cleanup
-func cleanupLoginItem(withPath path: String) {
-    print("Removing \(path) from login items...")
-
-    guard let items = LSSharedFileListCreate(nil, kLSSharedFileListSessionLoginItems.takeUnretainedValue(), nil)?.takeUnretainedValue() else {
-        print("[-] Could not obtain the list of Login Items.")
-        return
-    }
-
-    guard let loginItems = LSSharedFileListCopySnapshot(items, nil)?.takeRetainedValue() as? [LSSharedFileListItem] else {
-        print("[-] Could not copy the snapshot of Login Items.")
-        return
-    }
-
-    for item in loginItems {
-    var resolvedURL: Unmanaged<CFURL>?
-    LSSharedFileListItemResolve(item, 0, &resolvedURL, nil)
-        if let actualURL = resolvedURL?.takeRetainedValue() as URL?, actualURL.path == path {
-            if CFURLGetString(actualURL as CFURL) as String == path {
-                let result = LSSharedFileListItemRemove(items, item)
-                if result == noErr {
-                    print("[+] Cleanup: \(path) has been removed from Login Items")
-                } else {
-                    print("[-] Error removing \(path) from Login Items")
-                }
-            }
-        }
-    }
-}
-
-// Remove a file from the filesystem
-func removeFile(atPath path: String) {
-    do {
-        try FileManager.default.removeItem(atPath: path)
-        print("[+] Successfully removed file at: \(path)")
-    } catch {
-        print("[-] Failed to remove file at: \(path)")
-    }
-}
-
 // Main Execution
-print("This TTP will add the provided path or shell command to Login Items. If cleanup is true, it will then remove it.")
+print("This TTP will add the provided path or shell command to Login Items.")
 
 // Check if we have the necessary argument
-guard CommandLine.arguments.count > 2 else {
-    print("[!] Missing parameters. Usage: programName <path> <true|false>")
+guard CommandLine.arguments.count > 1 else {
+    print("[!] Missing parameters. Usage: programName <path>")
     exit(1)
 }
 
 let commandOrPath = CommandLine.arguments[1]
-let cleanupAction = CommandLine.arguments[2].lowercased()
 
-// Check for default /Users/Shared/calc.sh behavior
+// Execute default behavior if specified
 if commandOrPath == "/Users/Shared/calc.sh" {
     let script = """
     #!/bin/bash
@@ -62,12 +21,27 @@ if commandOrPath == "/Users/Shared/calc.sh" {
     """
 
     let fileMan = FileManager.default
-    let scriptData = Data(script.utf8)
-    try? fileMan.createFile(atPath: commandOrPath, contents: scriptData, attributes: nil)
 
-    var attributes = [FileAttributeKey: Any]()
-    attributes[.posixPermissions] = 0o755
-    try? fileMan.setAttributes(attributes, ofItemAtPath: commandOrPath)
+    // Check if /Users/Shared/ exists, if not, create it
+    if !fileMan.fileExists(atPath: "/Users/Shared/") {
+        do {
+            try fileMan.createDirectory(atPath: "/Users/Shared/", withIntermediateDirectories: true, attributes: nil)
+        } catch {
+            print("[-] Failed to create directory at /Users/Shared/. Error: \(error)")
+        }
+    }
+
+    // Create the file
+    do {
+        let scriptData = Data(script.utf8)
+        try fileMan.createFile(atPath: commandOrPath, contents: scriptData, attributes: nil)
+        var attributes = [FileAttributeKey: Any]()
+        attributes[.posixPermissions] = 0o755
+        try fileMan.setAttributes(attributes, ofItemAtPath: commandOrPath)
+        print("[+] Successfully created file at: \(commandOrPath)")
+    } catch {
+        print("[-] Failed to create file at: \(commandOrPath). Error: \(error)")
+    }
 }
 
 // Always add to login items first
@@ -93,14 +67,4 @@ do {
     print("[+] \(commandOrPath) successfully added as a Login Item")
 } catch let error {
     print("[-] Error: \(error)")
-}
-
-// If cleanup flag is true, remove the item from login items
-if cleanupAction == "true" {
-    cleanupLoginItem(withPath: commandOrPath)
-
-    // If it's the default calc script, remove it
-    if commandOrPath == "/Users/Shared/calc.sh" {
-        removeFile(atPath: commandOrPath)
-    }
 }
